+++ srcs/free/free_pipeline.c
@@ -6,7 +6,7 @@ void	free_pipeline(t_pipeline *pipeline)
 
 	vec_int_free(pipeline->args);
 	i = 0;
-	while (i < pipeline->execves->size)
+	while (i < (int)pipeline->execves->size)
 		i++;
 	free_execves(pipeline->execves);
 	free(pipeline);
